% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_depths.R
\name{classify_depths}
\alias{classify_depths}
\title{Classify bathymetry into depth zones}
\usage{
classify_depths(bathymetry_raster, planning_grid = NULL)
}
\arguments{
\item{bathymetry_raster}{}

\item{planning_grid}{}
}
\value{
A raster of depth zones, rasterized to the planning grid resolution and CRS if supplied.
}
\description{
Depth classifications are:
* 0 - 200m: Epipelagic zone
* 200 - 1000m: Mesopelagic zone
* 1000 - 4000m: Bathypelagic zone
* 4000 - 6000m: Abyssopelagic zone
* 6000m+: Hadopelagic zone
}
