% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bathymetry.R
\name{get_bathymetry}
\alias{get_bathymetry}
\title{Get bathymetry data for the area of interest}
\usage{
get_bathymetry(
  area_polygon,
  planning_grid = NULL,
  bathymetry_data_filepath = NULL,
  resolution = 1,
  keep = FALSE,
  path = NULL,
  download_timeout = 300
)
}
\arguments{
\item{area_polygon}{an sf polygon or multipolygon object of the area of interest (e.g., a country's EEZ)}

\item{planning_grid}{a raster template with the desired resolution and coordinate reference system generated by `get_planning_grid()`; values in areas of interest are 1, while all other values are NA}

\item{bathymetry_data_filepath}{string; the file path (including file name and extension) where bathymetry raster data are saved locally}

\item{resolution}{numeric; the resolution (in minutes) of data to pull from the ETOPO 2022 Global Relief model. Values less than 1 can only be 0.5 (30 arc seconds) and 0.25 (15 arc seconds)}

\item{keep}{logical; whether to save the bathymetry data locally}

\item{path}{string; the file path where you would like to save bathymetry data}

\item{download_timeout}{numeric; the maximum number of seconds a query is allowed to run}
}
\value{
A raster of bathymetry for the area of interest, rasterized to the planning grid resolution and CRS if supplied.
}
\description{
This function extracts bathymetry data for the area of interest from the ETOPO 2022 Global Relief model. If data are already downloaded locally, the user can specify the file path of the dataset. See Details for more information.
}
\details{
If the user has downloaded data for the area of interest, for example from GEBCO (https://www.gebco.net), they can pass the file path to this function and it will crop and rasterize the data using the supplied planning grid. If no file path is provided, the function will extract bathymetry data for the area from the ETOPO 2022 Global Relief model served by NOAA (https://www.ncei.noaa.gov/products/etopo-global-relief-model).
}
\examples{
# Grab EEZ data first 
bermuda_eez <- get_area(area_name = "Bermuda")
# Grab bathymetry data
bathymetry <- get_bathymetry(area_polygon = bermuda_eez)
}
