% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coral_habitat.R
\name{get_coral_habitat}
\alias{get_coral_habitat}
\title{Get coral habitat suitability data for area of interest}
\usage{
get_coral_habitat(
  area_polygon,
  planning_grid = NULL,
  antipatharia_threshold = 0.22,
  octocoral_threshold = 2
)
}
\arguments{
\item{area_polygon}{}

\item{planning_grid}{}

\item{antipatharia_threshold}{}

\item{octocoral_threshold}{}
}
\value{
A raster stack with of coral habitat suitabilities within the area polygon, this is rasterized using the planning_grid input raster if supplied
}
\description{
Antipatharia: Global extent of black coral habitat suitability modelled using Maxent. The antipatharia habitat suitability was converted to a presence/ absence map "by choosing a threshold value of habitat suitability based on the maximum sum of sensitivity and specificity (threshold mss = 0.23)". Text from the source paper, Yesson et al. 2017. This threshold can be altered via the function input.
}
\details{
Cold water corals: Modelled habitat suitability for five species of Scleractinia, using Maxent. Presence defined using Maxent values above the lowest 10% of values (see Fig 7 in reference)

Octocorals: Modelled global habitat suitability for 7 species of cold-water octocoral, found deeper than 50m. "A consensus/summary map incorporating all seven octocoral suborders was constructed by generating binary presence/absence maps from the model outputs indicating areas of high suitability using a score threshold that maximized the sum of the specificity and sensitivity based on the validation data (Carroll, 2010). These binary layers were summed to generate a layer containing the number of octocoral suborders predicted to be present per cell."
}
