% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_shore.R
\name{get_dist_shore}
\alias{get_dist_shore}
\title{Get distance to shore for a spatial grid}
\usage{
get_dist_shore(spatial_grid)
}
\arguments{
\item{spatial_grid}{a `terra::rast()` or `sf` planning grid with the desired resolution and coordinate reference system generated by `get_grid()`. For a raster, values in the planning grid cells should be 1, while all other values are NA}
}
\value{
a `terra::rast` or `sf` object (same type as `spatial_grid` input) with distance to shore for each grid cell.
}
\description{
Calculates distance from shore for each grid cell in the provided spatial grid. Spatial grid can be `terra::rast()` or `sf` format. The Natural Earth high resolution land polygons are used as the coastline and are downloaded from the Natural Earth website (https://www.naturalearthdata.com/downloads/10m-physical-vectors/), so an internet connection is required.
}
\examples{
# Get EEZ data first 
bermuda_eez <- get_area(area_name = "Bermuda", mregions_column = "territory1")
# Get a planning grid
planning_grid <- get_grid(area_polygon = bermuda_eez, projection_crs = '+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs', resolution = 5000)
dist_from_shore_rast <- get_dist_shore(planning_grid)
terra::plot(dist_from_shore_rast)
}
