% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecoregion.R
\name{get_ecoregion}
\alias{get_ecoregion}
\title{Get ecoregions for area of interest or a spatial grid}
\usage{
get_ecoregion(
  area_polygon = NULL,
  spatial_grid = NULL,
  type = "MEOW",
  antimeridian = NULL
)
}
\arguments{
\item{area_polygon}{an \code{sf} polygon or multipolygon object of the area of interest (e.g., a country's EEZ)}

\item{spatial_grid}{a \code{terra::rast()} or \code{sf} grid with the desired resolution and coordinate reference system generated by \code{get_grid()}. For a raster, values in the grid cells should be 1 for the study area and NA for all other cells.}

\item{type}{which ecoregion type is required? Default is \code{MEOW} (Marine Ecosystems of the World); other possible values are \code{Longhurst} and \verb{\\"LME\\"}}

\item{antimeridian}{does \code{area_polygon} or \code{spatial_grid} span the antimeridian? If so, this should be set to \code{TRUE}, otherwise set to \code{FALSE}. If set to \code{NULL} (default) the function will try to check if data spans the antimeridian and set this appropriately.}
}
\value{
If an \code{area_polygon} is supplied, an \code{sf} object of the chosen ecoregions for that area. If a \code{spatial_grid} is supplied, a raster or \code{sf} of gridded ecoregions data, depending on \code{spatial_grid} format.
}
\description{
Gets ecoregion data within an area polygon or planning grid depending on which is supplied
}
\details{
The Following ecoregions can be obtained:
\itemize{
\item Marine Ecosystems of the World \href{https://www.worldwildlife.org/publications/marine-ecoregions-of-the-world-a-bioregionalization-of-coastal-and-shelf-areas}{dataset}
}
}
\examples{
#' # Get EEZ data first 
bermuda_eez <- get_area(area_name = "Bermuda", mregions_column = "territory1")
# Get Marine Ecoregions of the World data
ecoregions <- get_ecoregion(area_polygon = bermuda_eez)
# Get Longhurst Provinces in a planning grid
bermuda_grid <- get_grid(area_polygon = bermuda_eez, projection_crs = '+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs', resolution = 5000)
longhurst_gridded <- get_ecoregion(spatial_grid = bermuda_grid, type = "Longhurst")
}
