% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eez.R
\name{get_eez}
\alias{get_eez}
\title{Get an EEZ for a country of interest}
\usage{
get_eez(country_name)
}
\arguments{
\item{country_name}{The name of the EEZ of interest}
}
\value{
The EEZ as an sf object
}
\description{
This uses the `mrp_get` function from the `mregions2` package to retrieve an EEZ object from Marine Regions (https://marineregions.org/gazetteer.php). For a full list of EEZs available, use `marineregions2::mrp_col_unique("eez", "territory1")`
}
\examples{
get_eez("Bermuda")
}
