% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enviro_regions.R
\name{get_enviro_regions}
\alias{get_enviro_regions}
\title{Create environmental regions for area of interest}
\usage{
get_enviro_regions(
  area_polygon,
  planning_grid = NULL,
  show_plots = TRUE,
  raw_data = FALSE,
  num_clusters = NULL,
  max_num_clusters = 8
)
}
\arguments{
\item{area_polygon}{}

\item{bo_layer_codes}{}

\item{data_dir}{}
}
\description{
Create environmental regions for area of interest
}
