% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_etopo_bathymetry.R
\name{get_etopo_bathymetry}
\alias{get_etopo_bathymetry}
\title{Get bathymetry data for the area of interest - helper function to `get_bathymetry()`}
\usage{
get_etopo_bathymetry(aoi, resolution, keep, path, download_timeout)
}
\arguments{
\item{aoi}{an sf polygon or multipolygon object of the area of interest (e.g., a country's EEZ)}

\item{resolution}{numeric; the resolution (in decimal degrees) of data to pull from the ETOPO 2022 Global Relief model}

\item{keep}{logical; whether to save the bathymetry data locally}

\item{path}{string; the file path where you would like to save bathymetry data}

\item{download_timeout}{numeric; the maximum number of seconds a query is allowed to run}
}
\value{
A raster of bathymetry for the area of interest
}
\description{
This function extracts bathymetry data for the area of interest from the ETOPO 2022 Global Relief model, using a script similar to that from `marmap::getNOAA.bathy()`
}
\keyword{internal}
