% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gfw.R
\name{get_gfw}
\alias{get_gfw}
\title{Get data from Global Fishing Watch}
\usage{
get_gfw(
  spatial_grid = NULL,
  raw = FALSE,
  resolution = "LOW",
  start_year = 2018,
  end_year = 2023,
  group_by = "location",
  summarise = "mean_total_annual_effort",
  key = gfwr::gfw_auth()
)
}
\arguments{
\item{spatial_grid}{\code{sf} or \code{terra::rast()} grid, e.g. created using
\code{get_grid()}. Alternatively, if raw data is required, an \code{sf} polygon can
be provided, e.g. created using \code{get_boundary()}, and set \code{raw = TRUE}.}

\item{raw}{\code{logical} if TRUE, \code{spatial_grid} can be an \code{sf} polygon, and the
raw GFW data, in tibble format, is returned for a bounding box covering the
polygon +/-1 degree. The bounding box is expanded to ensure that the entire
polygon has data coverage once the point data is rasterized. This data will
not be summarised, i.e. \code{summarise} is ignored.}

\item{resolution}{\code{string} either \code{"HIGH"} = 0.01 degree spatial resolution,
or \code{"LOW"}  = 0.1.}

\item{start_year}{\code{numeric} must be 2012 or more recent year. Note that GFW
added data from Spire and Orbcomm AIS providers in 2017, so data from 2017
is likely to have greater spatial and temporal coverage (\href{https://doi.org/10.1126/sciadv.abq2109}{Welch et al. 2022}).}

\item{end_year}{\code{numeric} any year between 2012 and the current year}

\item{group_by}{\code{string} can be \code{"geartype"}, \code{"flag"}, or \code{"location"}.}

\item{summarise}{\code{string} can be \code{"total_annual_effort"}; the sum of all
fishing effort for all years, from \code{start_year} to \code{end_year} at each
location is returned, grouped by \code{group_by}; or
\code{"mean_total_annual_effort"}; the mean of the annual sums of fishing effort
for all years at each location is returned, grouped by \code{group_by}.}

\item{key}{\code{string} Authorization token. Can be obtained with gfw_auth()
function. See \code{gfwr}
\href{https://github.com/GlobalFishingWatch/gfwr?tab=readme-ov-file#authorization}{website}
for details on how to request a token.}
}
\value{
For gridded data, a \code{terra::rast()} or \code{sf} object, depending on the
\code{spatial_grid} format. If \code{raw = TRUE}, non-summarised data in \code{tibble}
format is returned for the polygon area direct from the GFW query
\code{gfwr::get_raster()}.
}
\description{
Global Fishing Watch (GFW) provides data on apparent fishing
effort (hours) at 0.01 degree spatial resolution, based on automatic
identification system (AIS) broadcasts from vessels. Data is principally
for larger vessel (> 24m in length); less than 1\% of vessels <12 m length
are represented in the data (see \href{https://globalfishingwatch.org/dataset-and-code-fishing-effort/}{GFW website}
for detailed information). This function is primarily a wrapper for the
\href{https://github.com/GlobalFishingWatch/gfwr}{\code{gfwr} package} function
\code{get_raster()}, but allows the user to return multiple years of data in a
summarized and gridded format. An API key is required to retrieve GFW data;
see the package website for instructions on how to get and save one (free).
}
\examples{
\dontshow{if (nchar(Sys.getenv("GFW_TOKEN"))>0) withAutoprint(\{ # examplesIf}
#get mean total annual fishing effort for Bermuda for the years 2022-2023
#first get a grid for Bermuda
bermuda_grid <- get_boundary(name = "Bermuda") \%>\% get_grid(resolution = 0.1, crs = 4326)

bermuda_gfw_effort <- get_gfw(spatial_grid = bermuda_grid, start_year = 2022)

#plot the data
terra::plot(bermuda_gfw_effort)

#get total fishing effort for each gear type in Fiji's EEZ for 2022
fiji_grid <- get_boundary(name = "Fiji") \%>\% get_grid(resolution = 1e4, crs = "+proj=tcea +lon_0=178 +datum=WGS84 +units=m +no_defs", output = "sf_square")

fiji_gfw_effort <- get_gfw(spatial_grid = fiji_grid, start_year = 2022, end_year = 2022, group_by = "geartype", summarise = "total_annual_effort")

plot(fiji_gfw_effort, border = FALSE)

#quantile is better for viewing the fishing effort distribution due to the long tail of values
plot(fiji_gfw_effort[1], border= FALSE, breaks = "quantile")
\dontshow{\}) # examplesIf}
}
