% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seamount_peaks.R
\name{get_seamount}
\alias{get_seamount}
\title{Get seamounts}
\usage{
get_seamount(
  spatial_grid = NULL,
  raw = FALSE,
  buffer = NULL,
  name = "seamounts",
  antimeridian = NULL
)
}
\arguments{
\item{spatial_grid}{\code{sf} or \code{terra::rast()} grid, e.g. created using \code{get_grid()}. Alternatively, if raw data is required, an \code{sf} polygon can be provided, e.g. created using \code{get_boundary()}, and set \code{raw = TRUE}.}

\item{raw}{\code{logical} if TRUE, \code{spatial_grid} should be an \code{sf} polygon, and the raw seamounts data in that polygon(s) will be returned}

\item{buffer}{\code{numeric}; the distance from the seamount peak to include in the output. Distance should be in the same units as the spatial_grid, use e.g. \code{sf::st_crs(spatial_grid, parameters = TRUE)$units_gdal} to check units. If buffering raw data, this units are degrees.}

\item{name}{name of raster or column in sf object that is returned}

\item{antimeridian}{Does \code{spatial_grid} span the antimeridian? If so, this should be set to \code{TRUE}, otherwise set to \code{FALSE}. If set to \code{NULL} (default) the function will try to check if data spans the antimeridian and set this appropriately.}
}
\value{
If an \code{area_polygon} is supplied, an \code{sf} seamounts object. If a \code{spatial_grid} is supplied, a raster or \code{sf} of gridded seamount data, depending on \code{spatial_grid} format.
}
\description{
Get seamounts data in a spatial grid or polygon
}
\details{
Seamounts are classified as peaks at least 1000m higher than the surrounding seafloor \href{https://doi.org/10.3354/meps07268}{Morato et al. 2008}. The seamounts peak dataset is from \href{https://doi.org/10.14324/111.444/ucloe.000030}{Yeson et al. 2021}.
}
\examples{
# Get EEZ data first 
bermuda_eez <- get_boundary(name = "Bermuda")
# Get raw seamounts data
seamount_peaks <- get_seamount_peaks(spatial_grid = bermuda_eez, raw = TRUE)
# Get gridded seamount data
bermuda_grid <- get_grid(boundary = bermuda_eez, crs = '+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs', resolution = 1000)
seamounts_gridded <- get_seamount_peaks(spatial_grid = bermuda_grid)
}
