% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seamount_peaks.R
\name{get_seamount_peaks}
\alias{get_seamount_peaks}
\title{Get seamount peaks for area of interest}
\usage{
get_seamount_peaks(area_polygon)
}
\arguments{
\item{area_polygon}{}
}
\value{
An sf object of seamounts for the area of interest
}
\description{
Seamounts, classified as peaks at least 1000m higher than the surrounding seafloor [Morato et al. 2008](https://doi.org/10.3354/meps07268). These data are from [Yesson et al. 2021](https://doi.org/10.14324/111.444/ucloe.000030).
}
