% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seamounts.R
\name{get_seamounts}
\alias{get_seamounts}
\title{Get seamounts}
\usage{
get_seamounts(
  spatial_grid = NULL,
  raw = FALSE,
  buffer = NULL,
  name = "seamounts",
  antimeridian = NULL
)
}
\arguments{
\item{spatial_grid}{\code{sf} or \code{terra::rast()} grid, e.g. created using \code{get_grid()}. Alternatively, if raw data is required, an \code{sf} polygon can be provided, e.g. created using \code{get_boundary()}, and set \code{raw = TRUE}.}

\item{raw}{\code{logical} if TRUE, \code{spatial_grid} should be an \code{sf} polygon, and the raw seamounts data in that polygon(s) will be returned}

\item{buffer}{\code{numeric}; the distance from the seamount peak to include in the output. Distance should be in the same units as the spatial_grid, use e.g. \code{sf::st_crs(spatial_grid, parameters = TRUE)$units_gdal} to check units. If buffering raw data, units are metres, unless \code{sf::sf_use_s2()} is set to \code{FALSE}, in which case the units are degrees.}

\item{name}{name of raster or column in sf object that is returned}

\item{antimeridian}{Does \code{spatial_grid} span the antimeridian? If so, this should be set to \code{TRUE}, otherwise set to \code{FALSE}. If set to \code{NULL} (default) the function will try to check if data spans the antimeridian and set this appropriately.}
}
\value{
For buffered seamounts as gridded data, a \code{terra::rast()} or \code{sf} object, depending on the \code{spatial_grid} format. If \code{raw = TRUE} and \code{buffer = NULL} an \code{sf} POINT geometry object of seamount peaks within the polygon provided. If \code{raw = TRUE} and \code{buffer} is not \code{NULL} an \code{sf} polygon geometry object of buffered seamount peaks within the polygon provided. Note: at present, it is not possible to return gridded seamount peaks: https://github.com/emlab-ucsb/oceandatr/issues/48
}
\description{
Get seamounts data in a spatial grid or polygon
}
\details{
Seamounts are classified as peaks at least 1000m higher than the surrounding seafloor \href{https://doi.org/10.3354/meps07268}{Morato et al. 2008}. The seamounts peak dataset is from \href{https://doi.org/10.14324/111.444/ucloe.000030}{Yeson et al. 2021}.
\href{https://doi.org/10.1073/pnas.0910290107}{Morato et al. 2010} found that seamounts have higher biodiversity within 30 - 40 km of the peak. To enable this radius of higher biodiversity to be included in conservation planning, the \code{buffer} argument can be set, so that each seamount peak is buffered to the radius specified
}
\examples{
# Get EEZ data first 
bermuda_eez <- get_boundary(name = "Bermuda")
# Get raw seamounts data
seamount_peaks <- get_seamount_peaks(spatial_grid = bermuda_eez, raw = TRUE)
# Get gridded seamount data
bermuda_grid <- get_grid(boundary = bermuda_eez, crs = '+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs', resolution = 10000)
#buffer seamounts to a distance of 30 km (30,000 m)
seamounts_gridded <- get_seamount_peaks(spatial_grid = bermuda_grid, buffer = 30000)
}
