% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seamounts_buffered.R
\name{get_seamounts_buffered}
\alias{get_seamounts_buffered}
\title{Get seamount peak areas buffered to a specified radius}
\usage{
get_seamounts_buffered(area_polygon, planning_grid, buffer_km = 30)
}
\arguments{
\item{area_polygon}{}

\item{planning_grid}{}

\item{buffer_km}{}
}
\value{
A raster of buffered seamount areas found within the area polygon, rasterized using the planning_grid input raster
}
\description{
Seamounts, classified as peaks at least 1000m higher than the surrounding seafloor [Morato et al. 2008](https://doi.org/10.3354/meps07268). These data are from [Yesson et al. 2021](https://doi.org/10.14324/111.444/ucloe.000030). Each peak is buffered to the distance specified in the function call which has to be in kilometres
}
