% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceandatr-package.R
\docType{package}
\name{oceandatr-package}
\alias{oceandatr}
\alias{oceandatr-package}
\title{oceandatr: Ocean Data Acquisition and processing}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

For acquiring, processing, and gridding ocean related data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://emlab-ucsb.github.io/oceandatr/}
}

}
\author{
\strong{Maintainer}: Jason Flower \email{jflower@ucsb.edu} (\href{https://orcid.org/0000-0002-6731-8182}{ORCID})

Authors:
\itemize{
  \item Jason Flower \email{jflower@ucsb.edu} (\href{https://orcid.org/0000-0002-6731-8182}{ORCID})
  \item Echelle S. Burns \email{echelle_burns@ucsb.edu} (\href{https://orcid.org/0000-0002-3902-1410}{ORCID})
}

Other contributors:
\itemize{
  \item Jason D. Everett \email{JaseEverett@gmail.com} (\href{https://orcid.org/0000-0002-6681-8054}{ORCID}) [contributor]
}

}
\keyword{internal}
