% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offshoredatr-package.R
\docType{package}
\name{offshoredatr-package}
\alias{offshoredatr}
\alias{offshoredatr-package}
\title{offshoredatr: Offshore Data Preparation}
\description{
Offshore data preparation for use in prioritizations
}
\author{
\strong{Maintainer}: Jason Flower \email{jflower@ucsb.edu} (\href{https://orcid.org/0000-0002-6731-8182}{ORCID})

}
\keyword{internal}
