% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatialgridr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{get_data_in_grid}
\alias{get_grid}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{spatialgridr}{\code{\link[spatialgridr]{get_data_in_grid}}, \code{\link[spatialgridr]{get_grid}}}
}}

