% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_empty_layers}
\alias{remove_empty_layers}
\title{Remove empty layers in spatial object}
\usage{
remove_empty_layers(dat)
}
\arguments{
\item{dat}{\code{sf} or \code{terra::rast} object}
}
\value{
\code{sf} or \code{terra::rast} depending on input
}
\description{
Removes any layers (\code{terra::rast} object) or columns (\code{sf} object) that are all zero or NA
}
