% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_depths.R
\name{classify_depths}
\alias{classify_depths}
\title{Classify bathymetry data into depth zones}
\usage{
classify_depths(bathymetry_raster, planning_grid = NULL)
}
\arguments{
\item{bathymetry_raster}{a raster of bathymetry data (meters) that you wish to categorize; depths must be negative}

\item{planning_grid}{a raster template with the desired resolution and coordinate reference system generated by `get_planning_grid()`; values in areas of interest are 1, while all other values are NA}
}
\value{
A raster of depth zones, rasterized to the planning grid resolution and CRS if supplied.
}
\description{
This function takes a bathymetric raster (depths must be negative and in meters) and reclassifies depths to categorical depth zones. If geospatial data are not in the appropriate coordinate reference system, users can specify a planning grid to reproject the bathymetry data.
}
\details{
The depth groups are categorized as follows: 
\itemize{
\item Epipelagic Zone: 0-200 m depth
\item Mesopelagic Zone: 200-1000 m depth
\item Bathypelagic Zone: 1000-4000 m depth
\item Abyssopelagic Zone: 4000-6000 m depth
\item Hadopelagic Zone: 6000+ m depth
}
}
\examples{
# Grab EEZ data first 
bermuda_eez <- get_eez(country_name = "Bermuda")
# Grab bathymetry data
bathymetry <- get_bathymetry(area_polygon = bermuda_eez)
# Classify depths 
depth_zones <- classify_depths(bathymetry)
depth_zones
}
