% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area.R
\name{get_area}
\alias{get_area}
\title{Get an a shapefile for the area of interest}
\usage{
get_area(
  query_type = "eez",
  area_name,
  mregions_column = "territory1",
  show_options = FALSE
)
}
\arguments{
\item{query_type}{string; the area type that you would like to query for; some options include "eez", "high seas", "ecoregions" (default is "eez")}

\item{area_name}{string; the name of the country or area that you would like to query for; must match a name in `mregions2` for the `mregions_column` selected}

\item{mregions_column}{string; the name of the column in `mregions2` that you would like to query to find the area corresponding to `area_name` (defaut is "territory1")}

\item{show_options}{logical; whether to show all of the options available in `mregions2` or not; running this will not query `mregions2` (default is FALSE)}
}
\value{
if `show_options` = FALSE, returns the sf object of the area of interest; if `show_options` = TRUE, a named list will be returned. Names in the list correspond to options for `mregions_column` and values correspond to options for `area_name`
}
\description{
This uses the `mrp_get` function from the `mregions2` package to retrieve an area object from Marine Regions (https://marineregions.org/gazetteer.php). You can filter using any column available in the `mregions2` database. To see all possible column names and codes, specify `show_options` = TRUE when you run the function.
}
\examples{
# Show all possible EEZ areas
get_area(show_options = TRUE)
# Get EEZ area just for Bermuda
get_area(area_name = "Bermuda")
# Get EEZ areas for Kiribati (iso_ter1 = KIR)
get_area(area_name = "KIR", mregions_column = "iso_ter1")
}
