% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bathymetry.R
\name{get_bathymetry}
\alias{get_bathymetry}
\title{Get bathymetry data}
\usage{
get_bathymetry(
  spatial_grid = NULL,
  raw = FALSE,
  classify_bathymetry = TRUE,
  above_sea_level_isNA = FALSE,
  name = "bathymetry",
  bathymetry_data_filepath = NULL,
  path = NULL,
  antimeridian = NULL
)
}
\arguments{
\item{spatial_grid}{\code{sf} or \code{terra::rast()} grid, e.g. created using
\code{get_grid()}. Alternatively, if raw data is required, an \code{sf} polygon can
be provided, e.g. created using \code{get_boundary()}, and set \code{raw = TRUE}.}

\item{raw}{\code{logical} if TRUE, \code{spatial_grid} should be an \code{sf} polygon, and
the raw data in that polygon(s) will be returned}

\item{classify_bathymetry}{\code{logical}; whether to classify the bathymetry into
depth zones. Original bathymetry data can be classified if \code{raw = TRUE} and
\code{spatial_grid} is an \code{sf} polygon.}

\item{above_sea_level_isNA}{\code{logical}; whether to set bathymetry (elevation)
data values that are above sea level (i.e. greater than or equal to zero)
to \code{NA} (\code{TRUE}) or zero (\code{FALSE})}

\item{name}{\code{string}; name of raster or column in sf object that is returned}

\item{bathymetry_data_filepath}{\code{string}; the file path (including file name
and extension) where bathymetry raster data are saved locally}

\item{path}{\code{string}; the folder path where you would like to save the bathymetry
data. Defaults to \code{tempdir()}}

\item{antimeridian}{Does \code{spatial_grid} span the antimeridian? If so, this
should be set to \code{TRUE}, otherwise set to \code{FALSE}. If set to \code{NULL}
(default) the function will try to check if \code{spatial_grid} spans the
antimeridian and set this appropriately.}
}
\value{
If \code{classify_bathymetry = FALSE}, bathymetry data in the
\code{spatial_grid} supplied, or in the original raster file resolution if \code{raw = TRUE}. If \code{classify_bathymetry = TRUE} a multi-layer raster or an \code{sf}
object with one zone in each column is returned, depending on the
\code{spatial_grid} format. If \code{classify_bathymetry = TRUE} and \code{raw = TRUE} (in
which case \code{spatial_grid} should be an \code{sf} polygon), the raw raster
bathymetry data is classified into depth zones.
}
\description{
Get bathymetry data from the GEBCO 2025 global terrain model. If data are already downloaded locally, the user can specify the file path of the dataset. Data can be classified into depth zones by setting \code{classify_bathymetry = TRUE}
}
\details{
Extracts bathymetry data for an \code{area_polygon}, or if a
\code{spatial_grid} is supplied, gridded bathymetry is returned.

Data can be classified into depth zones by setting \code{classify_bathymetry = TRUE}. Depths are classified as follows:
\itemize{
\item Continental Shelf: 0 - 200 m depth
\item Upper Bathyal: 200 - 800 m depth
\item Lower Bathyal: 800 - 3500 m depth
\item Abyssal: 3500 - 6500 m depth
\item Hadal: 6500+ m depth
}

If the user has downloaded bathymetry data for the area of interest, for
example from GEBCO (https://www.gebco.net), they can pass the file path to
this function in \code{bathymetry_data_filepath}. If no file path is provided,
the function will extract bathymetry data for the area from the GEBCO 2025
global terrain model (sub-ice) from  the Natural Environment Research Council's (NERC) Centre for Environmental Data Analysis (CEDA)
(https://data.ceda.ac.uk/bodc/gebco/global/gebco_2025).
}
\examples{
# Get EEZ data first
bermuda_eez <- get_boundary(name = "Bermuda")
# Get raw bathymetry data, not classified into depth zones
bathymetry <- get_bathymetry(spatial_grid = bermuda_eez, raw = TRUE, classify_bathymetry = FALSE)
terra::plot(bathymetry)
# Get depth zones in spatial_grid
bermuda_grid <- get_grid(boundary = bermuda_eez, crs = '+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs', resolution = 20000)
depth_zones <- get_bathymetry(spatial_grid = bermuda_grid)
terra::plot(depth_zones)
#It is also possible to get the raw bathymetry data in gridded format by setting raw = FALSE and classify_bathymetry = FALSE
bermuda_grid_sf <- get_grid(boundary = bermuda_eez, crs = '+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs', resolution = 20000, output = "sf_hex")
gridded_bathymetry <- get_bathymetry(spatial_grid = bermuda_grid_sf, classify_bathymetry = FALSE)
plot(gridded_bathymetry)
}
