% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bathymetry.R
\name{get_bathymetry}
\alias{get_bathymetry}
\title{Get bathymetry for area of interest}
\usage{
get_bathymetry(
  area_polygon,
  planning_grid = NULL,
  bathymetry_data_filepath = NULL,
  resolution = 1,
  keep = TRUE,
  antimeridian = FALSE,
  path = "../data"
)
}
\arguments{
\item{area_polygon}{}

\item{planning_grid}{}

\item{bathymetry_data_filepath}{}

\item{resolution}{}

\item{keep}{}

\item{antimeridian}{}

\item{path}{}
}
\value{
A raster of bathymetry for the area of interest, rasterized to the planning grid resolution and CRS if supplied.
}
\description{
Best source of bathymetry data is GEBCO (https://www.gebco.net). If the user has downloaded data for the area of interest from GEBCO, they can pass the file path to this function and it will crop and rasterize the data using the supplied planning grid. If no file path is provided, the function will extract bathymetry data for the area from the ETOPO1 Global Relief model (https://www.ngdc.noaa.gov/mgg/global/relief/).
}
