% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bathymetry.R
\name{get_bathymetry}
\alias{get_bathymetry}
\title{Get bathymetry data}
\usage{
get_bathymetry(
  spatial_grid = NULL,
  raw = FALSE,
  classify_bathymetry = TRUE,
  above_sea_level_isNA = FALSE,
  name = "bathymetry",
  bathymetry_data_filepath = NULL,
  resolution = 1,
  keep = FALSE,
  path = NULL,
  download_timeout = 300,
  antimeridian = NULL
)
}
\arguments{
\item{spatial_grid}{\code{sf} or \code{terra::rast()} grid, e.g. created using \code{get_grid()}. Alternatively, if raw data is required, an \code{sf} polygon can be provided, e.g. created using \code{get_boundary()}, and set \code{raw = TRUE}.}

\item{raw}{\code{logical} if TRUE, \code{spatial_grid} should be an \code{sf} polygon, and the raw bathmetry data in that polygon(s) will be returned}

\item{classify_bathymetry}{\code{logical}; whether to classify the bathymetry into depth zones}

\item{above_sea_level_isNA}{\code{logical}; whether to set bathymetry (elevation) data values that are above sea level (i.e. greater than or equal to zero) to \code{NA} (\code{TRUE}) or zero (\code{FALSE})}

\item{name}{\code{string}; name of bathymetry raster or column in sf object that is returned}

\item{bathymetry_data_filepath}{\code{string}; the file path (including file name and extension) where bathymetry raster data are saved locally}

\item{resolution}{\code{numeric}; the resolution (in minutes) of data to pull from the ETOPO 2022 Global Relief model. Values less than 1 can only be 0.5 (30 arc seconds) and 0.25 (15 arc seconds)}

\item{keep}{\code{logical}; whether to save the bathymetry data locally}

\item{path}{\code{string}; the file path where you would like to save bathymetry data}

\item{download_timeout}{\code{numeric}; the maximum number of seconds a query to the NOAA website is allowed to run}

\item{antimeridian}{Does \code{spatial_grid} span the antimeridian? If so, this should be set to \code{TRUE}, otherwise set to \code{FALSE}. If set to \code{NULL} (default) the function will try to check if data spans the antimeridian and set this appropriately.}
}
\value{
If \code{classify_bathymetry = FALSE}, bathymetry data gridded using the \code{spatial_grid} supplied, or raw bathymetry data if \code{raw = TRUE}
If \code{classify_bathymetry = FALSE}, depth zones data in the \code{spatial_grid} supplied, or in the original raster file resolution if \code{raw = TRUE}. For gridded data, a multi-layer raster object or an \code{sf} object with one zone in each column is returned, depending on the \code{spatial_grid} format.
}
\description{
Get bathymetry data for an area from the ETOPO 2022 Global Relief model. If data are already downloaded locally, the user can specify the file path of the dataset. Data can be classified into depth zones by setting \code{classify_bathymetry = TRUE}
}
\details{
Extracts bathymetry data for an \code{area_polygon}, or if a \code{spatial_grid} is supplied, gridded bathymetry is returned.

#' Data can be classified into depth zones by setting \code{classify_bathymetry = TRUE}. Depths are classified as follows:
\itemize{
\item Epipelagic Zone: 0-200 m depth
\item Mesopelagic Zone: 200-1000 m depth
\item Bathypelagic Zone: 1000-4000 m depth
\item Abyssopelagic Zone: 4000-6000 m depth
\item Hadopelagic Zone: 6000+ m depth
}

If the user has downloaded bathymetry data for the area of interest, for example from GEBCO (https://www.gebco.net), they can pass the file path to this function in \code{bathymetry_data_filepath}. If no file path is provided, the function will extract bathymetry data for the area from the ETOPO 2022 Global Relief model served by NOAA (https://www.ncei.noaa.gov/products/etopo-global-relief-model).
}
\examples{
# Get EEZ data first 
bermuda_eez <- get_boundary(name = "Bermuda")
# Get raw bathymetry data, not classified into depth zones
bathymetry <- get_bathymetry(spatial_grid = bermuda_eez, raw = TRUE, classify_bathymetry = FALSE)
# Get depth zones in spatial_grid
bermuda_grid <- get_grid(boundary = bermuda_eez, crs = '+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs', resolution = 20000)
depth_zones <- get_bathymetry(spatial_grid = bermuda_grid)
terra::plot(depth_zones)
}
