% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coral_habitat.R
\name{get_coral_habitat}
\alias{get_coral_habitat}
\title{Get coral habitat suitability data for area of interest}
\usage{
get_coral_habitat(
  area_polygon,
  planning_grid = NULL,
  antipatharia_threshold = 22,
  octocoral_threshold = 2
)
}
\arguments{
\item{area_polygon}{an sf polygon or multipolygon object of the area of interest (e.g., a country's EEZ)}

\item{planning_grid}{a raster template with the desired resolution and coordinate reference system generated by `get_planning_grid()`; values in areas of interest are 1, while all other values are NA}

\item{antipatharia_threshold}{numeric between 0 and 100; the threshold value for habitat suitability for antipatharia corals to be considered present (default is 22, as defined in Yesson et al., 2017)}

\item{octocoral_threshold}{numeric between 0 and 7; the threshold value for how many species (of 7) should be predicted present in an area for octocorals to be considered present (default is 2)}
}
\value{
A raster stack with of coral habitat suitabilities within the area polygon, this is rasterized using the planning_grid input raster if supplied
}
\description{
This function extracts coral habitat suitability in the area of interest and re-defines areas as presence (1) or absence (0) for three groups of deep-water coral species (antipatharia, cold water corals, octocorals). See Details for more information on the species data available.
}
\details{
Habitat suitability data are available for the following species groups: 
\itemize{
\item Antipatharia: The global extent of black coral habiatat suitability modelled using Maxent. The antipatharia habitat suitability is converted to a presence/absence map "by choosing a threshold value of habitat suitability based on the maximum sum of sensitivity and specificity (threshold mss = 0.23)" (Yesson et al. 2017). This threshold can be altered via the function input. Data from Yesson et al. (2017).
\item Cold water corals: The global habitat suitability for five species of Scleractinia modelled using Maxent. Presence defined using Maxent values above the lowest 10 percent of values. Data from Davies and Guinotte (2011).
\item Octocorals: The global habitat suitability modelled for 7 species of cold-water octocoral found deeper than 50m. Data from and described in Yesson et al. (2012): "A consensus/summary map incorporating all seven octocoral suborders was constructed by generating binary presence/absence maps from the model outputs indicating areas of high suitability using a score threshold that maximized the sum of the specificity and sensitivity based on the validation data (Carroll, 2010). These binary layers were summed to generate a layer containing the number of octocoral suborders predicted to be present per cell."
}
}
\examples{
# Grab EEZ data first 
bermuda_eez <- get_area(area_name = "Bermuda")
# Get coral habitat
coral_habitat <- get_coral_habitat(area_polygon = bermuda_eez)
}
