% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eez.R
\name{get_eez}
\alias{get_eez}
\title{Get an EEZ for a country of interest}
\usage{
get_eez(country_name)
}
\arguments{
\item{country_name}{The name of the EEZ of interest}
}
\value{
The EEZ as an sf object
}
\description{
This uses the `mr_gaz_records_by_name` and `mr_gaz_record` functions from the `mregions2` package to retrieve an EEZ object from the Marine Gazeeter (https://marineregions.org/gazetteer.php). Note that the EEZ name is often different to the country name, e.g. use "Bermudian" to get Bermuda's EEZ; see https://marineregions.org/gazetteer.php?p=search for a list of EEZ names
}
\examples{
get_eez("bermudian")
}
