% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enviro_regions.R
\name{get_enviro_regions}
\alias{get_enviro_regions}
\title{Create environmental regions for area of interest}
\usage{
get_enviro_regions(
  area_polygon,
  planning_grid = NULL,
  show_plots = TRUE,
  raw_data = FALSE,
  num_clusters = NULL,
  max_num_clusters = 8
)
}
\arguments{
\item{area_polygon}{an sf polygon or multipolygon object of the area of interest (e.g., a country's EEZ)}

\item{planning_grid}{a raster template with the desired resolution and coordinate reference system generated by `get_planning_grid()`; values in areas of interest are 1, while all other values are NA}

\item{show_plots}{logical; whether to show a series of paneled boxplots for each environmental variable in each environmental region (default is TRUE)}

\item{raw_data}{logical; whether to only download the raw data and not proceed to calculating environmental regions (default is FALSE)}

\item{num_clusters}{numeric; the number environmental regions to calculate - to be used when a clustering algorithm is not necessary (default is NULL)}

\item{max_num_clusters}{numeric; the maximum number of environmental regions to try when using the clustering algorithm (default is 8)}
}
\value{
a raster stack with each layer representing a unique environmental region (if `raw_data == FALSE`); a raster stack with each layer representing a dataset pulled from Bio-Oracle (if `raw_data == TRUE`)
}
\description{
This function extracts environmental data from Bio-Oracle for the area of interest and calculates environmental regions using a clustering algorithm. Details are described in Magris et al. (2020):https://doi.org/10.1111/ddi.13183
}
\details{
Environmental variables that are extracted include: mean carbon phytoplankton biomass, mean chlorophyll concentration, mean dissolved oxygen concentration, mean nitrate concentration, mean pH, mean phosphate concentration, mean primary production, mean sea surface salinity, maximum sea surface temperature, mean sea surface temperature, mininum sea surface temperature, and mean silicate concentration.
}
\examples{
# Grab EEZ data first 
bermuda_eez <- get_area(area_name = "Bermuda")
# Just grab environmental data in the area of interest
get_enviro_regions(bermuda_eez, raw_data = TRUE)
# Calculate three environmental regions from the data
get_enviro_regions(bermuda_eez, num_clusters = 3)
}
