% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geomorphology.R
\name{get_geomorphology}
\alias{get_geomorphology}
\title{Get geomorphology for an area of interest}
\usage{
get_geomorphology(area_polygon, planning_grid = NULL)
}
\arguments{
\item{area_polygon}{an sf polygon or multipolygon object of the area of interest (e.g., a country's EEZ)}

\item{planning_grid}{a raster template with the desired resolution and coordinate reference system generated by `get_planning_grid()`; values in areas of interest are 1, while all other values are NA}
}
\value{
A raster stack with geomorphological features found within the area polygon rasterized using the planning_grid input raster, or a list of geomorphological features in sf format
}
\description{
This function gathers geomorpholgical deatures in the area of interest.
}
\details{
Geomorphological features are from the Harris et al. 2014 dataset, available at: https://www.bluehabitats.org
}
\examples{
# Grab EEZ data first 
bermuda_eez <- get_eez(country_name = "Bermuda")
# Get geomorphology
geomorphology <- get_geomorphology(area_polygon = bermuda_eez)
}
