% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geomorphology.R
\name{get_geomorphology}
\alias{get_geomorphology}
\title{Get geomorphology for an area of interest}
\usage{
get_geomorphology(
  area_polygon,
  planning_grid = NULL,
  raster_or_vector = "vector"
)
}
\arguments{
\item{area_polygon}{}

\item{planning_grid}{}

\item{raster_or_vector}{}
}
\value{
A raster stack with geomorphological features found within the area polygon rasterized using the planning_grid input raster, or a list of geomorphological features in sf format
}
\description{
Get geomorphology for an area of interest
}
