% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_knolls.R
\name{get_knolls}
\alias{get_knolls}
\title{Get knolls base areas for area of interest or planning grid}
\usage{
get_knolls(
  area_polygon = NULL,
  spatial_grid = NULL,
  name = "knolls",
  antimeridian = NULL
)
}
\arguments{
\item{area_polygon}{\code{sf} polygon or multipolygon object of the area of interest (e.g., a country's EEZ)}

\item{spatial_grid}{\code{terra::rast()} or \code{sf} planning grid with the desired resolution and coordinate reference system generated by \code{get_grid()}. For a raster, values in the planning grid cells should be 1, while all other values are NA}

\item{name}{name of raster or column in sf object that is returned}

\item{antimeridian}{\code{logical} does \code{area_polygon} or \code{spatial_grid} span the antimeridian? If so, this should be set to \code{TRUE}, otherwise set to \code{FALSE}. If set to \code{NULL} (default) the function will try to check if data spans the antimeridian and set this appropriately.}
}
\value{
If an \code{area_polygon} is supplied, an \code{sf} object of knolls for that area. If a \code{spatial_grid} is supplied, a raster or \code{sf} of gridded knolls data, depending on \code{spatial_grid} format.
}
\description{
This function gets knolls base area data within an area polygon or planning grid depending on which is supplied
}
\details{
Knolls are small seamounts; seamounts with peaks 200-1000 m higher than the surrounding seafloor \href{https://doi.org/10.3354/meps07268}{(Morato et al., 2008)}. The knolls base area data is from \href{https://doi.org/10.1016/j.dsr.2011.02.004}{Yesson et al. 2011}
}
\examples{
# Get EEZ data first 
bermuda_eez <- get_area(area_name = "Bermuda", mregions_column = "territory1")
# Get knolls
knolls <- get_knolls(area_polygon = bermuda_eez)
# Get knolls in a planning grid
planning_grid <- get_grid(area_polygon = bermuda_eez, projection_crs = '+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs', resolution = 5000)
knolls_gridded <- get_knolls(spatial_grid = planning_grid)
}
