% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_knolls.R
\name{get_knolls}
\alias{get_knolls}
\title{Get knolls base areas for area of interest}
\usage{
get_knolls(area_polygon, planning_grid = NULL)
}
\arguments{
\item{area_polygon}{}

\item{planning_grid}{}
}
\value{
A raster of knolls found within the area polygon, rasterized using the planning_grid input raster, or an sf object of knolls for the area of interest if no planning grid is supplied
}
\description{
Knolls are 'small' seamounts, classified as seamounts between 200 and 1000m higher than the surrounding seafloor [Morato et al. 2008](https://doi.org/10.3354/meps07268). These data are the knoll base area data from [Yesson et al. 2011](https://doi.org/10.1016/j.dsr.2011.02.004).
}
