% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_knolls.R
\name{get_knolls}
\alias{get_knolls}
\title{Get knolls base areas for area of interest}
\usage{
get_knolls(area_polygon, planning_grid = NULL)
}
\arguments{
\item{area_polygon}{an sf polygon or multipolygon object of the area of interest (e.g., a country's EEZ)}

\item{planning_grid}{a raster template with the desired resolution and coordinate reference system generated by `get_planning_grid()`; values in areas of interest are 1, while all other values are NA}
}
\value{
A raster of knolls found within the area polygon, rasterized using the planning_grid input raster, or an sf object of knolls for the area of interest if no planning grid is supplied
}
\description{
This function identifies areas of knolls (small seamounts, seamounts with peaks 200-1000 m higher than the surrounding seafloor (Morato et al., 2008))
}
\details{
These data are the knoll base area data from Yesson et al. (2011)
}
\examples{
# Grab EEZ data first 
bermuda_eez <- get_eez(country_name = "Bermuda")
# Get knolls
knolls <- get_knolls(area_polygon = bermuda_eez)
}
