% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_knolls.R
\name{get_knolls}
\alias{get_knolls}
\title{Get knolls base areas for area of interest or planning grid}
\usage{
get_knolls(
  area_polygon = NULL,
  planning_grid = NULL,
  name = "knolls",
  antimeridian = NULL
)
}
\arguments{
\item{area_polygon}{an `sf` polygon or multipolygon object of the area of interest (e.g., a country's EEZ)}

\item{planning_grid}{a `terra::rast()` or `sf` planning grid with the desired resolution and coordinate reference system generated by `get_planning_grid()`. For a raster, values in the planning grid cells should be 1, while all other values are NA}

\item{name}{name of raster or column in sf object that is returned}

\item{antimeridian}{does `area_polygon` or `planning_grid` span the antimeridian? If so, this should be set to `TRUE`, otherwise set to `FALSE`. If set to `NULL` (default) the function will try to check if data spans the antimeridian and set this appropriately.}
}
\value{
If an `area_polygon` is supplied, an `sf` object of knolls for that area. If a `planning_grid` is supplied, a raster or `sf` of gridded knolls data, depending on `planning_grid` format.
}
\description{
This function gets knolls data within an area polygon or planning grid depending on which is supplied
}
\details{
Knolls are small seamounts; seamounts with peaks 200-1000 m higher than the surrounding seafloor [(Morato et al., 2008)](https://doi.org/10.3354/meps07268). The knolls data is from [Yesson et al. 2011](https://doi.org/10.1016/j.dsr.2011.02.004)
}
\examples{
# Get EEZ data first 
bermuda_eez <- get_area(area_name = "Bermuda")
# Get knolls
knolls <- get_knolls(area_polygon = bermuda_eez)
Get knolls in a planning grid
planning_grid <- get_planning_grid(area_polygon = bermuda_eez, projection_crs = '+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs', resolution = 5000)
knolls_gridded <- get_knolls(planning_grid = planning_grid)
}
