% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_planning_grid.R
\name{get_planning_grid}
\alias{get_planning_grid}
\title{Create a planning grid raster for an area of interest}
\usage{
get_planning_grid(area_polygon, projection_crs, resolution_km = 5)
}
\arguments{
\item{area_polygon}{An sf object that will be rasterized}

\item{projection_crs}{A suitable crs for the area of interest}

\item{resolution_km}{The desired planning unit (grid cell) resolution in km}
}
\value{
A raster planning grid of the same resolution and crs provided
}
\description{
Create a planning grid raster for an area of interest
}
