% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seamount_peaks.R
\name{get_seamount_peaks}
\alias{get_seamount_peaks}
\title{Get seamount peaks for area of interest}
\usage{
get_seamount_peaks(area_polygon)
}
\arguments{
\item{area_polygon}{an sf polygon or multipolygon object of the area of interest (e.g., a country's EEZ)}
}
\value{
An sf object of seamounts for the area of interest
}
\description{
This function collects data for seamount peak locations within the area of interest. Seamounts are classified as peaks at least 1000m higher than the surrounding seafloor (Morato et al. 2008)
}
\examples{
# Grab EEZ data first 
bermuda_eez <- get_area(area_name = "Bermuda")
# Get seamount peak locations 
seamount_peaks <- get_seamount_peaks(bermuda_eez)
}
