% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seamount_peaks.R
\name{get_seamount_peaks}
\alias{get_seamount_peaks}
\title{Get seamount peaks for area of interest or planning grid}
\usage{
get_seamount_peaks(
  area_polygon = NULL,
  spatial_grid = NULL,
  name = "seamounts",
  antimeridian = NULL
)
}
\arguments{
\item{area_polygon}{an `sf` polygon or multipolygon object of the area of interest (e.g., a country's EEZ)}

\item{spatial_grid}{a `terra::rast()` or `sf` planning grid with the desired resolution and coordinate reference system generated by `get_grid()`. For a raster, values in the planning grid cells should be 1, while all other values are NA}

\item{name}{name to give to the raster or sf column}

\item{antimeridian}{Does `area_polygon` or `spatial_grid` span the antimeridian? If so, this should be set to `TRUE`, otherwise set to `FALSE`. If set to `NULL` (default) the function will try to check if data spans the antimeridian and set this appropriately.}
}
\value{
If an `area_polygon` is supplied, an `sf` seamounts object. If a `spatial_grid` is supplied, a raster or `sf` of gridded seamount data, depending on `spatial_grid` format.
}
\description{
This function gets seamount peaks within an area polygon or planning grid depending on which is supplied
}
\details{
Seamounts are classified as peaks at least 1000m higher than the surrounding seafloor [Morato et al. 2008](https://doi.org/10.3354/meps07268). The seamounts peak dataset is from [Yeson et al. 2021](https://doi.org/10.14324/111.444/ucloe.000030).
}
\examples{
# Get EEZ data first 
bermuda_eez <- get_area(area_name = "Bermuda", mregions_column = "territory1")
# Get seamount peaks
seamount_peaks <- get_seamount_peaks(area_polygon = bermuda_eez)
# Get seamounts in a planning grid
planning_grid <- get_grid(area_polygon = bermuda_eez, projection_crs = '+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs', resolution = 5000)
seamounts_gridded <- get_seamount_peaks(spatial_grid = planning_grid)
}
