% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seamounts_buffered.R
\name{get_seamounts_buffered}
\alias{get_seamounts_buffered}
\title{Create spatial dataset of seamount peaks buffered to a specified radius for an area of interest or planning grid}
\usage{
get_seamounts_buffered(
  area_polygon = NULL,
  spatial_grid = NULL,
  buffer = NULL,
  name = "seamounts",
  antimeridian = NULL
)
}
\arguments{
\item{area_polygon}{an `sf` polygon or multipolygon object of the area of interest (e.g., a country's EEZ)}

\item{spatial_grid}{a `terra::rast()` or `sf` planning grid with the desired resolution and coordinate reference system generated by `get_grid()`. For a raster, values in the planning grid cells should be 1, while all other values are NA}

\item{buffer}{numeric; the distance from the seamount peak to include in the output. Distance should be in the same units as the area_polygon or spatial_grid provided, use e.g. `sf::st_crs(spatial_grid, parameters = TRUE)$units_gdal` to check what units your planning grid or area polygon is in (works for raster as well as sf objects)}

\item{name}{name to give to the raster or sf column}

\item{antimeridian}{Does `area_polygon` or `spatial_grid` span the antimeridian? If so, this should be set to `TRUE`, otherwise set to `FALSE`. If set to `NULL` (default) the function will try to check if data spans the antimeridian and set this appropriately.}
}
\value{
If an `area_polygon` is supplied, an `sf` buffered seamounts object. If a `spatial_grid` is supplied, a raster or `sf` of gridded buffered seamount data, depending on `spatial_grid` format.
}
\description{
This function gets seamount peak locations within the area of interest or planning grid and adds a specified buffer around each peak.
}
\details{
Seamounts are classified as peaks at least 1000m higher than the surrounding seafloor [Morato et al. 2008](https://doi.org/10.3354/meps07268). The seamounts peak dataset is from [Yeson et al. 2021](https://doi.org/10.14324/111.444/ucloe.000030).
[Morato et al. 2010](https://doi.org/10.1073/pnas.0910290107) find that seamounts have higher biodiversity within 30 - 40 km of the peak.
}
\examples{
# Grab EEZ data first 
bermuda_eez <- get_area(area_name = "Bermuda", mregions_column = "territory1")
# Get buffered seamounts in Bermuda's EEZ - warns that buffering is not correct for long/lat (WGS 84) data. Buffer is in decimal degrees
seamounts_buffered <- get_seamounts_buffered(area_polygon = bermuda_eez, buffer = 0.5)
# Get seamounts buffered to 30 km, in a planning grid
planning_grid <- get_grid(area_polygon = bermuda_eez, projection_crs = '+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs', resolution = 5000)
seamounts_buffered_gridded <- get_seamounts_buffered(spatial_grid = planning_grid, buffer = 30000)
}
