% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceandatr-package.R
\docType{package}
\name{oceandatr-package}
\alias{oceandatr}
\alias{oceandatr-package}
\title{oceandatr: Ocean Data Access}
\description{
Prepares ocean related datasets ready for use in spatial prioritization, i.e. data in gridded format.
}
\author{
\strong{Maintainer}: Jason Flower \email{jflower@ucsb.edu} (\href{https://orcid.org/0000-0002-6731-8182}{ORCID})

}
\keyword{internal}
